refac1

+++ utils/postgresConnector.js
@@ -1,12 +1,16 @@
 const associate = require("../models/associate");
+const { DEV, TEST } = require("../statics/envs");
 const dataSeeder = require("./dataSeeder");
 const db = require("./db")
+const dotenv = require('dotenv').config();
 
 module.exports = async () => {
   try {
     await associate();
     await db.sequelize.authenticate();
-    await db.sequelize.sync({ force: true });
+    if ([DEV, TEST].includes(process.env.NODE_ENV)) {
+      await db.sequelize.sync({ force: true });
+    }
     await dataSeeder();
     return true;
   } catch (err) {
